+++ packages/clipboard-rs/build.rs
@@ -1,0 +1,6 @@
+fn main() {
+    napi_build_3_setup();
+}
+
+#[allow(unused_imports)]
+use napi_build::setup as napi_build_3_setup;